.emacs: Add more F9 bindings
"F9 +" is this way because F9 is just above "]"
(which is to the right of "+"),
so this keeps the relative positions of the 2 keys in the binding
the same as on a ThinkPad keyboard
(but the F9 key's absolute position is farther to the right).

"F9 g" would be better for magit-status,
but it's already taken by previous-buffer.