Konsole: Map Control-Backspace to Alt-Backspace
Change the "Default (XFree 4)" key binding profile
to send the code for Alt-Backspace when Control-Backspace is pressed.

The default was to send the code for Control-W,
which closes the current buffer in console-mode ErgoEmacs.
Alt-Backspace does not have such a conflict, but still
deletes the previous word in zsh, bash and console emacs by default.

Note that "Backspace-Ctrl" is left unmodified from the default
to prevent the Backspace binding from conflicting
with the Backspace+Ctrl one.